requirements.txt
llama-index==0.12.30
llama-index-vector-stores-faiss==0.3.0
llama-index-llms-openai==0.3.33
python-dotenv==1.0.0
nltk==3.9.1
torch==2.6.0
transformers==4.51.2
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.3
openai==1.72.0
pillow==10.2.0
requests==2.31.0 